:art: move wdio configs and update docs
