+++ README.md
@@ -1,1 +1,29 @@
 # gtr-test-cucumber
+
+The cucumber example of testing a guitar practice app.
+
+## Setup
+
+This project uses node 6x and npm 3x.  We recommend using NVM (for Mac) to manage
+you node version. The most compatible version will be in he .nvmrc
+
+```
+nvm use
+npm install
+```
+
+Next you will need to install selenium:
+```
+npm run selenium-install
+```
+Start your local selenium server:
+```
+npm run selenium
+```
+
+### Running Tests
+
+```
+npm run test-local-dev
+```
+
