+++ package.json
@@ -18,18 +18,17 @@
   "license": "MIT",
   "main": "config/master.conf.js",
   "scripts": {
-    "build": "babel pages/src --presets babel-preset-es2015 --out-dir dist",
     "cucumber": "cucumberjs test/features/ -r test/stepDefinitions",
     "debug-local-dev": "DEBUG=true node  --harmony wdio test/config/local/dev.conf.js",
     "lint": "eslint ./*",
     "node-inspector": "node-inspector --debug-port 5859 --no-preload",
     "selenium": "selenium-standalone start",
     "selenium-install": "selenium-standalone install",
-    "test": "wdio config/jenkins/regressionSuite.travis.conf.js",
-    "test-local": "wdio config/local/suite.conf.js",
-    "test-local-dev": "wdio config/local/dev.conf.js",
-    "test-remote": "wdio config/sauceLabs/suite.sauce.conf.js",
-    "test-remote-dev": "wdio config/sauceLabs/dev.sauce.conf.js"
+    "test": "wdio wdioConfig/jenkins/regressionSuite.travis.conf.js",
+    "test-local": "wdio wdioConfig/local/suite.conf.js",
+    "test-local-dev": "wdio wdioConfig/local/dev.conf.js",
+    "test-remote": "wdio wdioConfig/sauceLabs/suite.sauce.conf.js",
+    "test-remote-dev": "wdio wdioConfig/sauceLabs/dev.sauce.conf.js"
   },
   "dependencies": {
     "babel-preset-es2015": "6.14.0",
@@ -43,7 +42,6 @@
     "request": "^2.74.0",
     "wdio-cucumber-framework": "0.2.5",
     "wdio-dot-reporter": "0.0.6",
-    "wdio-json-reporter": "0.0.4",
     "wdio-junit-reporter": "^0.1.0",
     "wdio-sauce-service": "^0.2.5",
     "wdio-spec-reporter": "0.0.3",
