+++ wdioConfig/master.conf.js
@@ -26,7 +26,6 @@ exports.config = {
         source: false,       // <boolean> hide source uris
         profile: [],        // <string[]> (name) specify the profile to use
         strict: false,      // <boolean> fail if there are any undefined or pending steps
-        tags: [], // ['@only', '@isolate'], // <string[]> (expression) only execute the features or scenarios with tags matching the expression
         timeout: 60000,     // <number> timeout for step definitions
         ignoreUndefinedDefinitions: true, // <boolean> Enable this config to treat undefined definitions as warnings.
         require: ['./stepDefinitions/'], // relative path to the step files
@@ -38,43 +37,34 @@ exports.config = {
         chai.expect();
     },
     afterStep: function (stepResult) { //eslint-disable-line object-shorthand
-        /**
-         * Assign Step Staus of Passed/Failed/Skipped to a variable.
-         */
+
+        /** Assign Step Status of Passed/Failed/Skipped to a variable. */
         status = stepResult.getStatus();
-        /**
-         * If the step passes or is skipped, skip browser console logging.
-         */
+
+        /** If the step passes or is skipped, skip browser console logging. */
         if (status !== `failed`) {
             return;
         }
-        /**
-         * Assign current step data to variables for error logging purposes
-         */
+        /** Assign current step data to variables for error logging purposes */
         feature = stepResult.getStep().getScenario().getFeature().getName();
         scenario = stepResult.getStep().getScenario().getName();
         step = stepResult.getStep().getName();
-        /**
-         * gets the current browser logs from selenium and assigns them to a variable
-         */
+
+        /** gets the current browser logs from selenium and assigns them to a variable */
         const logs = browser.log('browser');
-        /**
-         * Filters out everything except error messages and adds step/scenario/feature info to the errorLogs array
-         */
+
+        /** Filters out everything except error messages and adds step/scenario/feature info to the errorLogs array */
         logs.value.forEach(log => {
-            /**
-             * only print error messages
-             */
+
+            /** only print error messages */
             if (!log || typeof log.level !== 'string' || log.level.toLowerCase() !== 'severe') {
                 return;
             }
-            /**
-             * Assign time error was incurred to a variable
-             */
+
+            /** Assign time error was incurred to a variable */
             timestamp = log.timestamp.toString();
-            /**
-             * Add The test path and browser console message to the errorLogs array
-             */
+
+            /** Add The test path and browser console message to the errorLogs array */
             errorLogs.push({
                 name: `${status}: ${feature} ${scenario} ${step}`,
                 message: log.message,
@@ -84,9 +74,7 @@ exports.config = {
     },
     after: function (_, capabilities) { //eslint-disable-line object-shorthand
         /* eslint-disable no-console */
-        /**
-         * Skip if there are no severe browser errors to report
-         */
+        /** Skip if there are no severe browser errors to report */
         if (status === 'passed') {
             return;
         } else if ((status === 'failed' || status === 'skipped') && errorLogs.length === 0) {
